Add test split to coco

+++ datasets/deep_fruits.py
@@ -17,7 +17,7 @@ def parse_command_line():
 
 
 def deep_fruits(root):
-    data_type = ['train']
+    data_type = ['train', 'test']
     for data in data_type:
         json_dict = {"images": [], "annotations": [], "categories": []}
         categories = {}
@@ -41,34 +41,35 @@ def deep_fruits(root):
                     annotation = annotation.split()
 
                     file_name = annotation[0].split('/')[-1]
-                    image = Image.open(image_dir + file_name)
-                    json_dict['images'].append({
-                        'file_name': file_name,
-                        'height': image.size[1],
-                        'width': image.size[0],
-                        'id': image_id
-                    })
-                    os.rename(image_dir + file_name, image_coco_dir + file_name)
-
-                    bboxs = np.split(np.array(annotation[2:]), int(annotation[1]))
-                    for bbox in bboxs:
-                        x_min, y_min, x_max, y_max = list(map(int, bbox[:4]))
-
-                        if category not in categories:
-                            categories[category] = len(categories) + 1
-                        category_id = categories[category]
-
-                        json_dict['annotations'].append({
-                            "bbox": [x_min, y_min, x_max - x_min, y_max - y_min],
-                            "area": (x_max - x_min) * (y_max - y_min),
-                            "segmentation": [],
-                            "iscrowd": 0,
-                            "image_id": image_id,
-                            "category_id": category_id,
-                            "id": bbox_id
+                    if os.path.exists(image_dir + file_name):
+                        image = Image.open(image_dir + file_name)
+                        json_dict['images'].append({
+                            'file_name': file_name,
+                            'height': image.size[1],
+                            'width': image.size[0],
+                            'id': image_id
                         })
+                        os.rename(image_dir + file_name, image_coco_dir + file_name)
 
-                        bbox_id += 1
+                        bboxs = np.split(np.array(annotation[2:]), int(annotation[1]))
+                        for bbox in bboxs:
+                            x_min, y_min, x_max, y_max = list(map(int, bbox[:4]))
+
+                            if category not in categories:
+                                categories[category] = len(categories) + 1
+                            category_id = categories[category]
+
+                            json_dict['annotations'].append({
+                                "bbox": [x_min, y_min, x_max - x_min, y_max - y_min],
+                                "area": (x_max - x_min) * (y_max - y_min),
+                                "segmentation": [],
+                                "iscrowd": 0,
+                                "image_id": image_id,
+                                "category_id": category_id,
+                                "id": bbox_id
+                            })
+
+                            bbox_id += 1
                 image_id += 1
 
             json_dict['categories'].append({
